+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/scripts/atlas/DataSource.ts
@@ -1,6 +1,21 @@
 import * as atlas from "azure-maps-control"
 
 export class DataSource {
+    public static clear(map: atlas.Map, sources: atlas.source.DataSource[] | string[]) {
+        sources.forEach((src) => {
+            let ds: atlas.source.DataSource;
+
+            if (src instanceof atlas.source.DataSource) {
+                ds = src;
+            }
+            else {
+                ds = map.sources.getById(src) as atlas.source.DataSource;
+            }
+
+            ds.clear();
+        });
+    }
+
     public static async importDataFromUrl(map: atlas.Map, source: any, url: string)  {
         let ds: atlas.source.DataSource;